// starter-react-master/src/modules/app/components/routes/types.ts
export enum Paths {
  root = '/',
  login = '/login',
  register = '/register',
  confirmEmail = '/confirm-email/:token',
  dummyComponent = '/dummy',
  serviceDemo = '/service',
  forgotPassword = '/forgot-password',
  resetPassword = '/reset-password/:code',
  changePassword = '/change-password',
}